ServerGlobal/requirements.txt
blinker==1.9.0
cffi==2.0.0
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
Flask==3.1.2
flask-cors==6.0.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mutagen==1.47.0
mysql-connector-python==9.5.0
pillow==12.1.0
pycparser==3.0
PyJWT==2.8.0
python-dotenv==1.0.1
Werkzeug==3.1.5
